+++ src/main/java/com/yunsseong/ai_ex_server/auth/service/MemberDetailsService.java
@@ -1,6 +1,7 @@
 package com.yunsseong.ai_ex_server.auth.service;
 
 import com.yunsseong.ai_ex_server.member.application.MemberService;
+import com.yunsseong.ai_ex_server.member.domain.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -15,6 +16,6 @@ public class MemberDetailsService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return memberService.findByEmail(username);
+        return new CustomUserDetails(memberService.findByEmail(username));
     }
 }
